pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
SQLAlchemy = "^1.4.39"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
google-api-python-client = "^2.52.0"
oauth2client = "^4.1.3"
python-dotenv = "^0.20.0"
aiosqlite = "^0.17.0"
pytest = "^7.1.2"
httpx = "^0.23.0"
trio = "^0.21.0"
pytest-asyncio = "^0.18.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start"
